fix: cast to proper dtype in EmbeddingParallel

+++ src/transformers/integrations/tensor_parallel.py
@@ -974,7 +974,7 @@ class EmbeddingParallel(TensorParallelLayer):
             input_mask = mod._input_mask
             # Use multiplication instead of in-place assignment to preserve gradients
             mask_expanded = input_mask.unsqueeze(-1).expand_as(outputs)
-            outputs = outputs * (~mask_expanded).float()
+            outputs = outputs * (~mask_expanded).to(outputs.dtype)
             del mod._input_mask
 
         return all_reduce_forward(outputs, device_mesh)
